Move time module
