+++ firmware/foundation/src/application/mod.rs
@@ -1,3 +1,2 @@
 pub mod channels;
 pub mod state;
-pub mod time;
