+++ firmware/foundation/src/lib.rs
@@ -12,6 +12,7 @@ pub mod layout;
 pub mod midi;
 pub mod protocol;
 pub mod storage;
+pub mod time;
 
 /// A trait for types that can be converted to and from another type `T`
 pub trait Convertible<T>: Sized {
